+++ release_notes.md
@@ -5,7 +5,10 @@
 * Fixed various crashes when using subtables. The crash will occur when the first column
   of the subtable if of type `col_type_Timestamp` or if it is nullable and of type Bool, Int
   or OldDateTime. Caused by bad static `get_size_from_ref()` methods of columns. (#2101)
-
+* TimestampColumn::is_nullable() mismatched its real nulability. Also, when adding a new
+  Float/Double column to a Table with rowcount > 0, the new entries would be non-null
+  even though the column was created as nullable.
+  
 ### Breaking changes
 
 * Refactored the `SharedGroup` constructors and open methods to use a new